requirement.txt
absl-py==2.1.0
accelerate==0.13.2
aiohttp-cors==0.7.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==10.0.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.7.0
boto3==1.24.96
botocore==1.27.96
brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.0.4
cloudpickle==3.0.0
colorama==0.4.6
colorful==0.5.6
colorlog==6.8.0
contourpy==1.2.0
cryptography==39.0.0
cycler==0.12.1
cython==0.29.32
cytoolz==0.12.2
dask==2023.10.1
dataclasses==0.8
datasets==2.6.1
debugpy==1.8.0
decorator==5.1.1
diffusers==0.11.1
dill==0.3.5.1
distlib==0.3.8
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
egl-probe==1.0.2
einops==0.7.0
entrypoints==0.4
etils==1.5.2
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.15
filelock==3.13.1
fonttools==4.47.2
free-mujoco-py==2.1.6
frozenlist==1.4.1
fsspec==2023.12.2
fvcore==0.1.5.post20221221
gdown==5.0.0
gitdb==4.0.11
gitpython==3.1.41
glfw==1.12.0
google-api-core==2.15.0
google-auth-oauthlib==0.4.6
google-auth==2.26.2
googleapis-common-protos==1.62.0
gpustat==1.1.1
grpcio==1.47.1
gym==0.21.0
h5py==3.7.0
huggingface-hub==0.20.2
hydra-colorlog==1.2.0
hydra-core==1.2.0
idna==3.6
imagecodecs==2022.9.26
imageio-ffmpeg==0.4.7
imageio==2.22.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.16.2
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter-client==8.6.0
jupyter-core==5.7.1
kiwisolver==1.4.5
labmaze==1.0.6
llvmlite==0.36.0
locket==1.0.0
lxml==5.1.0
markdown==3.5.2
markupsafe==2.1.4
matplotlib-inline==0.1.6
matplotlib==3.6.1
monotonic==1.6
msgpack==1.0.7
mujoco==3.1.1
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.53.1
numcodecs==0.10.2
numpy==1.23.3
nvidia-ml-py==12.535.133
oauthlib==3.2.2
omegaconf==2.2.3
opencensus-context==0.1.3
opencensus==0.11.4
opencv-python==4.6.0
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
parso==0.8.3
partd==1.4.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
pip==22.2.2
platformdirs==4.1.0
ply==3.11
portalocker==2.8.2
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.42
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==9.0.0
pyasn1-modules==0.3.0
pyasn1==0.5.1
pybullet-svl==3.1.6.4
pycparser==2.21
pydantic-core==2.14.6
pydantic==2.5.3
pygame==2.1.2
pygments==2.17.2
pyjwt==2.8.0
pymunk==6.2.1
pyopengl==3.1.7
pyopenssl==23.2.0
pyparsing==3.1.1
pyqt5-sip==12.11.0
pyqt5==5.15.7
pysocks==1.7.1
python-dateutil==2.8.2
pytorch3d==0.7.0
pytorchvideo==0.1.5
pytz==2023.3.post1
pyu2f==0.1.5
pywavelets==1.4.1
pyyaml==6.0.1
pyzmq==25.1.2
r3m==0.0.0
ray==2.2.0
referencing==0.32.1
regex==2023.12.25
requests-oauthlib==1.3.1
requests==2.31.0
responses==0.18.0
robomimic==0.2.0
robosuite==1.2.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.6.2
scikit-image==0.19.3
scikit-video==1.1.11
scipy==1.9.1
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==69.0.3
shapely==1.8.4
shortuuid==1.0.11
sip==6.7.12
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.2
tabulate==0.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.1
tensorboardx==2.5.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.14.1
typing-extensions==4.9.0
tzdata==2023.4
unicodedata2==15.1.0
urllib3==1.26.18
virtualenv==20.25.0
wandb==0.13.3
wcwidth==0.2.13
werkzeug==3.0.1
wheel==0.42.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zarr==2.12.0
zipp==3.17.0